# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "AuthorityPayload": {
        "additionalProperties": false,
        "description": "What the assistant may do on somebody's behalf. Everything opt-in.",
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/Capability"
            },
            "title": "Capabilities",
            "type": "array"
          }
        },
        "title": "AuthorityPayload",
        "type": "object"
      },
...
      "DevicePayload": {
        "additionalProperties": false,
        "description": "One device, as its platform identifies it.",
        "properties": {
          "platform": {
            "$ref": "#/components/schemas/DevicePlatform"
          },
          "token": {
            "maxLength": 512,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "platform",
          "token"
        ],
        "title": "DevicePayload",
        "type": "object"
      },
      "DevicePlatform": {
        "description": "Which delivery path a device needs.",
        "enum": [
          "ios",
          "android"
        ],
        "title": "DevicePlatform",
        "type": "string"
      },
      "EscalationContextResponse": {
        "description": "An escalation as the app shows it: the untrimmed notification words and their fields.",
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "call_id": {
            "title": "Call Id",
            "type": "string"
          },
          "caller": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caller"
          },
          "caller_label": {
            "title": "Caller Label",
            "type": "string"
          },
          "delivery": {
            "$ref": "#/components/schemas/NotificationDelivery"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "established": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Established"
          },
          "needed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Needed"
          },
          "raised_at": {
            "format": "date-time",
            "title": "Raised At",
            "type": "string"
          },
          "reason": {
            "$ref": "#/components/schemas/EscalationReason"
          },
          "status": {
            "$ref": "#/components/schemas/EscalationStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "call_id",
          "status",
          "reason",
          "title",
          "caller_label",
          "body",
          "caller",
          "established",
          "needed",
          "raised_at",
          "ended_at",
          "delivery"
        ],
        "title": "EscalationContextResponse",
        "type": "object"
      },
      "EscalationReason": {
        "description": "Why the assistant wants the user.",
        "enum": [
          "caller_asked_for_the_user",
          "action_not_authorised",
          "decision_needs_the_user",
          "important_enough_to_interrupt",
          "cannot_understand_the_caller",
          "user_rule_requires_it"
        ],
        "title": "EscalationReason",
        "type": "string"
      },
      "EscalationStatus": {
        "description": "Whether the call the escalation belongs to is still going.",
        "enum": [
          "live",
          "ended"
        ],
        "title": "EscalationStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HoursPayload": {
        "additionalProperties": false,
        "description": "When the assistant answers; outside it, calls ring the user. `null` is always (D-030).",
        "properties": {
          "active": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimeWindowPayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "HoursPayload",
        "type": "object"
      },
...
      "NotificationDelivery": {
        "description": "What became of telling the user, surfaced in the app rather than raised anywhere.",
        "enum": [
          "pending",
          "delivered",
          "failed",
          "no_devices"
        ],
        "title": "NotificationDelivery",
        "type": "string"
      },
...
      "OnboardingResponse": {
        "description": "Where somebody is in the flow, listing only the steps this deployment asks, in order.",
        "properties": {
          "completed": {
            "items": {
              "$ref": "#/components/schemas/OnboardingStep"
            },
            "title": "Completed",
            "type": "array"
          },
          "is_complete": {
            "title": "Is Complete",
            "type": "boolean"
          },
          "next_step": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OnboardingStep"
              },
              {
                "type": "null"
              }
            ]
          },
          "remaining": {
            "items": {
              "$ref": "#/components/schemas/OnboardingStep"
            },
            "title": "Remaining",
            "type": "array"
          },
          "skipped": {
            "items": {
              "$ref": "#/components/schemas/OnboardingStep"
            },
            "title": "Skipped",
            "type": "array"
          }
        },
        "required": [
          "completed",
          "skipped",
          "remaining",
          "next_step",
          "is_complete"
        ],
        "title": "OnboardingResponse",
        "type": "object"
      },
      "OnboardingStep": {
        "description": "One screen of setup, asking about one group of preferences.",
        "enum": [
          "call_handling",
          "call_forwarding",
          "hours",
          "authority",
          "notifications"
        ],
        "title": "OnboardingStep",
        "type": "string"
      },
      "OnboardingUpdate": {
        "additionalProperties": false,
        "properties": {
          "skipped": {
            "default": false,
            "title": "Skipped",
            "type": "boolean"
          },
          "step": {
            "$ref": "#/components/schemas/OnboardingStep"
          }
        },
        "required": [
          "step"
        ],
        "title": "OnboardingUpdate",
        "type": "object"
      },
...
      "PreferencesResponse": {
        "description": "Everything, as it stands.",
        "properties": {
          "authority": {
            "$ref": "#/components/schemas/AuthorityPayload"
          },
          "call_handling": {
            "$ref": "#/components/schemas/CallHandlingPayload"
          },
          "hours": {
            "$ref": "#/components/schemas/HoursPayload"
          },
          "important_contacts": {
            "items": {
              "$ref": "#/components/schemas/ImportantContactPayload"
            },
            "title": "Important Contacts",
            "type": "array"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "notifications": {
            "$ref": "#/components/schemas/NotificationsPayload"
          },
          "personality": {
            "$ref": "#/components/schemas/PersonalityPayload"
          },
          "privacy": {
            "$ref": "#/components/schemas/PrivacyResponse"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "locale",
          "call_handling",
          "important_contacts",
          "hours",
          "authority",
          "notifications",
          "personality",
          "privacy"
        ],
        "title": "PreferencesResponse",
        "type": "object"
      },
      "PreferencesUpdate": {
        "additionalProperties": false,
        "description": "A change to some of it; an omitted section is left exactly as it was.",
        "properties": {
          "authority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthorityPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "call_handling": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CallHandlingPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "hours": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HoursPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "important_contacts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ImportantContactPayload"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Important Contacts"
          },
          "locale": {
            "anyOf": [
              {
                "maxLength": 16,
                "minLength": 2,
                "pattern": "^[A-Za-z]{2,3}(-[A-Za-z0-9]{2,8})*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "notifications": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NotificationsPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "personality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonalityPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "privacy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrivacyPayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "PreferencesUpdate",
        "type": "object"
      },
...
      "PrivacyResponse": {
        "description": "What is kept, and for how long, as stored and unbounded above.",
        "properties": {
          "transcript_retention_days": {
            "title": "Transcript Retention Days",
            "type": "integer"
          }
        },
        "required": [
          "transcript_retention_days"
        ],
        "title": "PrivacyResponse",
        "type": "object"
      },
...
      "RegisterDeviceRequest": {
        "additionalProperties": false,
        "description": "This device's current push token, and the one the platform rotated away from if known.",
        "properties": {
          "platform": {
            "$ref": "#/components/schemas/DevicePlatform"
          },
          "previous_token": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9._:-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Token"
          },
          "token": {
            "maxLength": 512,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "platform",
          "token"
        ],
        "title": "RegisterDeviceRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "An agent that handles phone calls on someone's behalf.",
    "title": "LetMeHandle",
    "version": "2026.9.13"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/devices": {
      "put": {
        "description": "Record this device's push token for the signed-in user, idempotently.",
        "operationId": "register_device_v1_devices_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register this device for escalation notifications",
        "tags": [
          "escalation"
        ]
      }
    },
    "/v1/devices/unregister": {
      "post": {
        "description": "Forget this device for the signed-in user; a POST so the token stays out of access logs.",
        "operationId": "unregister_device_v1_devices_unregister_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevicePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stop sending escalation notifications to this device",
        "tags": [
          "escalation"
        ]
      }
    },
    "/v1/escalations/{call_id}": {
      "get": {
        "description": "What the user was, or would have been, told about this escalation.",
        "operationId": "read_escalation_v1_escalations__call_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "call_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Call Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscalationContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The context of an escalation",
        "tags": [
          "escalation"
        ]
      }
    },
...
    "/v1/onboarding": {
      "get": {
        "operationId": "read_onboarding_v1_onboarding_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Where setup is",
        "tags": [
          "preferences"
        ]
      },
      "post": {
        "description": "Record a step as answered or skipped; a step this deployment does not ask is refused.",
        "operationId": "record_onboarding_step_v1_onboarding_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record a step",
        "tags": [
          "preferences"
        ]
      }
    },
    "/v1/preferences": {
      "get": {
        "operationId": "read_preferences_v1_preferences_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read preferences",
        "tags": [
          "preferences"
        ]
      },
      "patch": {
        "description": "Change the sections that were sent and leave the rest exactly as they were (D-023).",
        "operationId": "update_preferences_v1_preferences_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change some of it",
        "tags": [
          "preferences"
        ]
      },
      "put": {
        "description": "Set everything the request mentions, starting from the defaults.",
        "operationId": "replace_preferences_v1_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Replace preferences",
        "tags": [
          "preferences"
        ]
      }
    },
...
  }
}